summarization/requirements.txt
aniso8601==8.0.0
bert-extractive-summarizer==0.3.0
blis==0.2.4
boto3==1.11.1
botocore==1.14.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cymem==2.0.3
docutils==0.15.2
Flask==1.1.1
flask-restplus==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==2.6.0
MarkupSafe==1.1.1
murmurhash==1.0.2
neuralcoref==4.0
numpy==1.18.1
plac==0.9.6
preshed==2.0.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.1.8
requests==2.22.0
s3transfer==0.3.0
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
sentencepiece==0.1.85
six==1.13.0
spacy==2.1.3
srsly==1.0.1
thinc==7.0.8
torch==1.3.1
tqdm==4.41.1
transformers==2.3.0
urllib3==1.25.7
wasabi==0.6.0
Werkzeug==0.16.0
